Remove ability to access element from willInsertElement
While this was a useful feature, it was not reliable. In many cases,
the element was accessible, but it was not accessible for children of a
MetamorphView. This caused unexpected behavior when views expected to be
able to access the element and were unable to do so. Since this feature
could not be relied upon we opted to remove it with a strong warning so
users would know what was going on. In most cases, using
didInsertElement should work in place of willInsertElement.