If TextFieldSupport cannot find its element in DOM, do not try to update the value property.
This happens when a text field has focus and it or
one of its parent views is destroyed. The view
receives a focusOut event but the text field
is removed from DOM.

We could not reliably test this as the focusOut
event seems to ONLY be sent when you MANUALLY
CLICK on the field. Calling focus() on the element
does not make it receive a focusOut event.

If you can figure out a way to get this to work
in Chrome, please fix this jsfiddle:

http://jsfiddle.net/zgSnB/2/