Records move to default transaction after commit
If a transaction is committed, the records that
belong to it should be moved back to the default
transaction. Additionally, if the default
transaction is committed, those records should be
moved to the *new* default transaction.

Additionally, records should be able to be moved
between transactions if they are still clean.
Records can now be removed from transactions
with the remove() method.