Avoid ping-ponging bindings.
This solves the long-standing issue where a
computed property that returned an object would
trigger a computed property in the reverse
direction.

Initially, we hoped to avoid scheduling the
second binding at all, but because setting
properties could trigger observers, which could
themselves trigger more `set`s, this is the best
solution, and avoids a dangerous superfluous
`get`.