Fix bug in attributes from previous commit
We should always expect that values set on a
record's attribute round-trip through the
transform process at least once.

For example, if I set the number 1 on a string
attribute, I would expect that the next time
I retrieve that attribute, it is the string "1".

After that, however, subsequent gets should return
objects with the same identity.