Improve perf of object model affecting View
Previously, it was possible for large classes
to adversely impact the performance of instance
creation of subclasses.

This specifically affected classes whose
subclasses were instantiated before they were.

In short, all of the parent class' properties
were copied onto each instance of the subclass,
instead of applying them to the superclass'
prototype chain and remembering that they were
already applied.

Major props to @ebryn and @kselden for sitting
through an insane debugging session figuring
this out!