Add introspection for cached values
Adds the ability to query an object for the cached
value of a computed property. This is useful for
instances where you want to retrieve the value of
a lazily computed property if it exists, but do
not want to accidentally trigger its creation if
it has not been used yet.