Allow views to pass controller object to templates
To support the notion that views/templates have
an associated controller, have views have a
controller property that is passed along
recursively to view instances.

If a child view specifies a controller property,
the inherited controller will be overridden and
the new controller will propagate down the view
hierarchy.

This lays the groundwork for adding a `controller`
keyword to Handlebars templates. More commits to
follow.