Fix failing Handlebars test that relies on states
A previous commit introduced a test that tests
the integration between Handlebars and Ember's
state manager. Unfortunately, ember-handlebars
does not have a dependency on ember-states, so the
test only passes when all of the tests are run
at once. (This is the reason the failure was
not caught by Travis.)

This fix adds ember-states as a dependency.
Ideally, we could create a new testing-only
package that includes both handlebars and states
as a dependency and contains just the integration
test.