Add property defaultApplication to Ember namespace
This property returns the default application. If more applications are created, the first one is set to the default application. A newly created application with the flag `isDefaultApplication` set to true overwrites the default application. If a default application is destroyed, the next available application is set to the default. If there is no application available, Ember.defaultApplication is undefined.

This commit is the base for rewrite of Ember.View#append, as discussed in #668