Remove new deleted records from record arrays
Fixes #168. Previously, only records that were
sent to the adapter were being removed from
record arrays. This is because the logic for
removing them was in the deleted.start enter
transition event, but new records that are deleted
transition directly to the deleted.saved state
(because deleting a new record is effectively a
no-op; the adapter should never be notified about
the record).

As part of this change, we now have records
that transition from created.uncommitted directly
move themselves to their transaction's clean
bucket (as they do not need to be sent to the
adapter).

This also fixes a bug where updated and
uncommitted records were moving themselves to the
clean bucket. Only newly created records should
do this; updated records should move to the
transaction's deleted bucket when deleted.