Update Ember.View#append so it respects Ember.defaultApplication
Ember.View#append now accepts a parameter `application`. If an application is passed, the view is appended to the `rootElement` of the passed application. If no application is passed, the `rootElement` of the default application - set as property `defaultApplication` on Ember namespace - is used. If no application is available, the view is appended to document.body.