Notify parentView of childView changes for virtual views
Previously, if you had a standard view with a virtual child, we returned
and then cached the results of the virtual view's child views. However,
if the virtual view's child views change, the root level view is not
notified of this change and misreports the childViews. This commit
corrects that issue.