Rename routable state managers to Ember.Router
Previously, routing functionality was mixed into
every state manager.

To make clear that an application's primary
state manager is intended to help with
routing concerns, the `route` method has been
separated out into the `Ember.Router` class,
which you should use for the primary state
manager that controls the application's URL.

We believe this naming change re-enforces the
role, and makes it more clear to new developers
what is happening.