Fix failing unit tests due to deferred teardown
Object destruction is deferred until the end of
the run loop. This was causing failing unit
tests in some cases where the old event dispatcher
was not being torn down before a new test started.

This ensures that application creation and
teardown happen inside run loops. This commit
also starts the process of ensuring that no code
gets scheduled to run from an auto timer.

In the future we should throw an exception if
deferred code is run outside a run loop.