Add urlFor to Router
Note: In general, you will not want to use
urlFor outside of the Router. Instead you will
either set the appropriate URL on a controller to
be used in a view, or use a feature we will write
next to allow a URL to be extracted from an event
that triggers a transition.