QUnit test runner: Run all packages by default
The test `if (!packages.length)` didn't actually do anything before,
because split always returns at least ['']. But it's more convenient
anyway to just run everything by default, rather than displaying an
error message. This way people can open the test runner without
parameters and it Just Works.

Note on the diff: The if main clause is deleted, and the rest is just a
dedent.