URL generation takes into account location type
Currently, when generating a URL for a given
router's event (as used in the {{action}} helper),
an absolute URL is generated.

This is problematic when, for example, using the
hash-based location object.

In this commit, we have the router give the
location object an opportunity to reformat the URL
before it is placed into the DOM as a link's
`href` attribute.