Call preventDefault on events handled through {{action}}
The idea is that we'd like to be able to use arbitrary view methods as
action handlers. As it is, they need to `return false` or call
`e.preventDefault()` to be usable as handlers.

For the (presumably few) cases where default handling is desired, we can
still register our own handlers/delegates with jQuery, or use event
handling on views or with eventManagers, which does not have automatic
preventDefault/stopPropagation.