Call setupStateManager in initialize, not in didBecomeReady
`initialize` needs to be called anyway in order to use routing, and as
it is, it requires being called before didBecomeReady is triggered,
since didBecomeReady checks if App.stateManager is defined.
This can introduce subtle ordering bugs or race conditions.

Moving all the router initialization code in one place solves these
issues.

Also rename setupStateManager to startRouting to be more descriptive.

This also addresses #911 and #889.