Injections can specify the order they are run.
Now, injections can register not just a function,
but also a name, and, optionally, the name of
the injection they should be run before.

This is useful in the case of, e.g., the data
store, which should create the instance of
DS.Store before the controllers are instantiated,
as they may have a dependency on the store at
initialization time.