Use evented system for dom events on views
Now we are using real events for all dom events on views. We check if
the callback is defined or if listeners are set. If you use callbacks,
you can return `false` in order to prevent default
In order to achive it goal, this commit adds `has` method to Evented
mixin.