A route's `serialize` should handle null contexts
In particular, when we are unrouting (in
anticipating of routing due to a URL change), the
default `unroute` implementation calls
`transitionTo` to transition to the parent, but
without a context. In these cases, if the parent
state has a context as part of its URL, it will
try to serialize this empty context, which causes
an error.

This naïve solution just serializes null contexts
to the empty string. Since we are unrouting and
the serialized value will never be used, this
resolves the problem. The long-term fix is to
have a smarter solution than just unrouting.