Change connectOutlet API to prefer Strings
Instead of:

  connectOutlet(App.PostView)

Do:

  connectOutlet('post')

The usage of a view before was hiding the way that
the controller was looked up. Reverse engineering
a string and getting a controller from the String
was confusing.

Now, you pass a string and we determine both the
view class and the controller to use.

You can override the details via:

  connectOutlet({
    viewClass: App.PostView,
    controller: postController,
    context: post
  })

You can specify an alternate outlet name:

  connectOutlet({
    outletName: 'detail',
    name: 'post'    
    context: post
  })

You can specify either a name or a view class,
but not both.