Add `isLoaded` flag for ManyArray
This commit refactors the internals to make it
easier to know when all records belonging to a
ManyArray have finished loading.

Now, the store keeps a map of all clientIds that
are being loaded and are members of a ManyArray.

Also, the ManyArray state manager has a new
initial `loading` state that it exits when all
of its clientIds are loaded into the store.