Do not run functions passed to Ember.assert, Ember.warn, and Ember.deprecate
This implicit behavior seems like a bit of a code smell, and might lead
to subtle bugs. Grepping through the Ember and ember-data sources
doesn't reveal any places where it is relied upon.

If this is ever needed again, I would suggest implementing it as a
separate set of functions.

I'm not changing the boolean casts that are everywhere, like
Ember.assert('...', !!foo), since they don't hurt and we don't want to
invite merge conflicts.