Sort routes by segment to detect nested paths
This improves the sorting algorithm used to find a route match.  Instead
of sorting the route as one string, this sorts on each segment of the
route using the following algorithm:

    * Prefer more segments
    * Prefer static segment over dynamic segment
    * Prefer segment with longer length
    * Sort alphabetically ascending to ensure deterministic search

This results in routes being sorted like:

    /products/:product_id/services/:id
    /products/:id
    /users/

Thus allowing non-nested states with nested paths to be correctly
detected and routed in routePath.