Change {{action}} API for more explicit contexts
Previously, the action helper automatically sent
the current context as the context to the event
it was triggering.

This meant that the router could receive
superfluous contexts that it would try to apply
to a parent state.

Instead of:

  {{action edit context="post"}}

Do:

  {{action edit post}}

If you were relying on the default context,
change:

  {{action edit}}

to:

  {{action edit this}}

This also paves the way for support for multiple
contexts.