Add FilterableMixin
Works in a similar way to the SortableMixin.

Mix this into an ArrayProxy, define the filterProperties and the array
will be filtered to only include the objects whose filterProperties are
truthy.

You can get more advanced filtering by defining a custom
filterCondition. The filterCondition is used to determine if the object is
to be included in the filtered list.  Any item properties used in
the filterCondition will need to be lised in filterProperties.