Make ApplicationView available on App.applicationView
This is analogous to App.applicationController.

The motivation for having App.applicationView is that you may have to
destroy the running app in teardown methods in your test suite, but you
don't want to destroy the entire app with App.destroy() (because then
you cannot start it up again).

This also means that we don't have to rely on private instance variables
anymore (_createdApplicationView, _createdRouter) when determining what
to destroy. Instead we use App.get('applicationView') and
App.get('router'). This is consistent with App.get('eventDispatcher').