Allow extension of chosen prototypes instead of the current all or none
One of my current projects involves embedding an ember widget into
a rather large pre-existing application   Extending the Array prototype
causes a slew of issues for both first party and third party code
in use on the relevant pages.

This change is backwards compatible and enables, via updating the
handling of the ENV.EXTEND_PROTOTYPES variable, the ability to choose
exactly which prototypes to extend:

window.Env = {
    EXTEND_PROTOTYPES: {
        Function: true,
        String: true,
        Array: true
    }
};

If window.ENV.EXTEND_PROTOTYPES is true or undefined
Ember.ENV.EXTEND_PROTOTYPES is initialized to the above hash.

If false it is initialized to the above hash with false intead of
true set for each prototype.

Otherwise it is assuemd to be a hash configured properly for the
application at hand.