Fix State.transitionTo documentation by removing @scope
With the `@scope`, the documentation would come out as a single garbled
(unformatted) paragraph.

Now it is `@scope`'d to Ember.State.prototype instead of Ember.State,
which I guess is not 100% correct, but we still have the `@static` tag
to tell it from an instance method. And at least it works. ;-)