Add support unhandledEvent for unhandledEvent to StateManager.
If an action is unhandled by the current state or any of its
parent states, the StateManager will now look for a method
named `unhandledEvent` in the current state, traversing up
through its parents until one is found. The method is called
with the manager, original event name and the context
argument(s) originally sent with the event. There is a default
implemenation of `unhandledEvent` on the StateManager itself,
which raises an Ember.Error unless `errorOnUnhandledEvent` is
false.

This flexibility may be useful for a variety of situations, but
it is specifically laying the groundwork for improved error
messages in Ember Data's DS.StateManager.

Luke Melia & Trek Glowacki