Resolves an issue when re-rendering MetamorphViews
Previously, it was possible for a view to get
into a bad state if it was re-rendered multiple
times in the same run loop. This lead some users
to experience deprecation warnings about
multiple re-renders.

This commit fixes the bug by ensuring that we
clear rendered children and re-render at the
same time that we move MetamorphViews into the
inDOM state.

Additionally, the deprecation warning has been
removed and instead now raise an error.