Ember.Select shouldn't trigger change for when content is an empty array-like obejct The previous check for triggering when content was empty relied on the falsiness of `[]` but would incorrectly trigger change when the content was an Array-like object whose proxied `length` was 0.
Specifically this would cause issues with setting the `contentBinding`
of a Select to the `DS.Model.find` values that are empty initially
setting the Select's `selection` from whatever the current value is to
`undefined` when change was triggered on an empty content.

Asynchronously, data would return and the content would become populated
but the selection value had already been unset.