App.container was not meant to be a public API
It appears that people are starting to use
App.container.lookup as the public API for
globally getting other controllers outside of the
router or other controllers.

This was not intended to be a public API.

Instead, you should trigger events on a route
(or a parent route), which have access to
`this.controllerFor`.

Example:

    App.ApplicationRoute.extend({
      events: {
        search: function(term) {
          this.controllerFor('search')
            .set('term', term);
        }
      }
    });

    App.SearchField = Ember.TextField.extend({
      insertNewline: function() {
        var term = this.get('value'),
            controller = this.get('controller');

        controller.send('search', term);
        this.set('value', '');
      }
    });

Events sent to the controller will bubble up to
the current route, and then up the route hierarchy