Reimplemented App.router as per documentation
I know the router API is in flux, and the documentation probably hasn't been updated for a while, however I've added this back in as it is incredibly useful during development to be able to access the router in the console.

Updated App.router implementation to have router set the application router property

Conditionally set App.router if not already set, in case multiple routers are in use.