Fix to observe nested property at `Ember.TextField`
There is a case that the nested property observing doesn't work on
`Ember.TextField`.
This case is that the content of observing property is updated *without value property*
(In this case, `Ember.TextArea` works.)

I think that the nested property should be observed not only `Ember.TextArea` but
also `Ember.TextField`.