Added readOnly flag to computed.alias.
Aliasing properties is powerful, to some peoples
surprise they alias in both direction (set and get).
Usually this is the behaviour one wants but sometimes,
as
https://github.com/emberjs/data/commit/7b88a92c090dddf2d554fb4d4095322f9eef8a59
suggests, this behaviour
is not always wanted.

When the read only flag is flipped on, a set will
throw and Error. Preventing the set, and providing
useful feedback.