Check for not supported events for {{action}} helper
This commit introduces a check for allowed event types on an {{action}}
helper. By default the events 'change', 'click', 'contextMenu',
'doubleClick', 'input', 'mouseDown', 'mouseUp' and 'submit' can be handled
with the {{action}} helper. Using an other eventType such as 'touchStart'
raises an error stating that this type of event shall be handled on a custom
Ember.View.

The list of allowed action event types can be extended by overwriting the
`customActionEvents` array on the Ember.Application:

```
App = Ember.Application.create({
  customActionEvents: ['focusIn', 'focusOut']
});
```