Enhance Ember.TargetActionSupport and introduce Ember.ViewTargetActionSupport
Enhancements to `Ember.TargetActionSupport` now allow `triggerAction` to receive an object specifying the desired action name and target. A target and/or action provided in the function call take precedence over any target/action properties in the object.

`Ember.ViewTargetActionSupport` is a new mixin that encompasses `Ember.TargetActionSupport` and complements it by defining the default `target` as the view's `controller` and the default `actionContext` as the view's `context`.