{{view}} helper performs container lookup
If the context lookup fails to resolve to a View class or instance, the
`view` helper will perform a lookup on the container. Accepted syntax:

    {{view widget}}
    {{view 'view:widget'}}

These two will ultimately perform a container lookup for "view:widget",
though the first will only do so if a `widget` lookup on the present
helper context fails to resolve to a View class or instance. The second
one is obviously more precise. Unfortunately, the quotes are required by
Handlebars.

The rationale for all of this is to bring a container-based approach to
templates and allow for things like type injections to work for views,
which can't work if a global path to the view class is provided, which
causes it to be initialized outside of the container.