Make parentController available from an itemController
When using the `{{#each}}` helper with itemController specified,
the itemController instances will now have the parentController
set on them. The parentController is the value of the `controller`
keyword when the #each helper executes.

If ArrayController is used directly with the itemController property
set, the itemController instances will get the ArrayController instance
as their parentController property.