Added `mergedProperties` to ember-metal, Ember.Route's `events`
/cc @lukemelia, @kselden

Added `mergedProperties`, the object-merging sibling of
`concatenatedProperties`. Properties specified in the
`mergedProperties` array will not overwrite but rather
merge into such properties defined in extended parent
classes or mixins. Merged methods can call _super

Also added it to the `events` hash in `Ember.Route`,
where it is most useful. Will add it to
Controllers per #2884 in a separate PR once
this is merged.