Fix the jQuery patch code for ember-testing click in Firefox.
The first problem was that Firefox was not firing all the expected events on checkboxes that were not in the DOM, so we have to make sure the checkboxes are appended to the body before testing them. The second problem was that Firefox fires change/click events on checkboxes in a different order than everybody else [1], so we can't make the more exact assertion about events fired without doing browser sniffing or more in depth bug sniffing, which didn't quite seem worth it.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=843554