Immediate return on null/undefined arguments for pushObjects method (issue #2967)
Ember.MutableArray and Ember.ArrayProxy method #pushObjects should accept only
Ember.Enumerable or Array as argument. In other cases it throws an exception.

However, if the equivalent of nothing (i.e. null/undefined) is passed to #pushObjects, then no exception should be thrown and the method should return immediately.

This is a bugfix for: #2967

Pull request contains: bugfix and tests.