Improve error message for missing itemView
This commit adds infrastructure for a container to return a
human-readable description for a fullName.

For example, Ember's default container will return App.FooController for
the fullName "controller:foo" (where App is the name of the namespaces).

This paves the way for improving other assertions to describe the
precise name of the class that Ember is looking for, rather than
container keys (which don't match the mental model of an Ember user).

Alternate resolvers should implement lookupDescription(fullName), and
Ember.Application will automatically wire that method up to the
container's `describe` method.