Ember.controllerFor and Route#controllerFor no longer generate controllers
In the standard use case, Route#controllerFor is called to get the controller
for another route. This only works reliably when the controller has already
been instantiated and had its model set by the associated route. However, it
was possible to call Route#controllerFor for a route that had not yet been
entered. In this case, a new controller would be generated, but without the
associated model. This also could lead to the wrong type of controller being
created. This commit resolves that already broken case.