linkTo param of 0 should be treated as a url param
Part of this fix is in router.js, the other part in Ember.

`linkTo` was generating incorrect hrefs when the provided
parameter was the number 0 (even though '0') worked. Now it
will work as expected without requiring casting to a string
first.

Also removed some unnecessary code on Ember's side to check that all
bound params are truthy before reporting that `routeArgs` changed in
`LinkView`.